+++ packages/vtable/src/core/BaseTable.ts
@@ -37,7 +37,7 @@ import {
   type ITableThemeDefine,
   InteractionState
 } from '../ts-types';
-import type { ColumnIconOption } from '../ts-types';
+import type { AnyFunction, ColumnIconOption } from '../ts-types';
 import { event, style as utilStyle } from '../tools/helper';
 
 import { TABLE_EVENT_TYPE } from './TABLE_EVENT_TYPE';
@@ -149,9 +149,10 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
   headerStyleCache: any;
   bodyStyleCache: any;
+  bodyBottomStyleCache: any;
   container: HTMLElement;
   isReleased: boolean = false;
-
+  _chartEventMap: Record<string, { query?: any; callback: AnyFunction }[]> = {};
   constructor(container: HTMLElement, options: BaseTableConstructorOptions = {}) {
     super();
     if (!container && options.mode !== 'node') {
@@ -359,6 +360,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
     this.headerStyleCache = new Map();
     this.bodyStyleCache = new Map();
+    this.bodyBottomStyleCache = new Map();
 
     internalProps.stick = { changedCells: [] };
 
@@ -2937,7 +2939,12 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     } else {
       cacheKey = row;
     }
-    let cacheStyle = this.bodyStyleCache.get(cacheKey);
+    let cacheStyle;
+    if (layoutMap.isBottomFrozenRow(row)) {
+      cacheStyle = this.bodyBottomStyleCache.get(cacheKey);
+    } else {
+      cacheStyle = this.bodyStyleCache.get(cacheKey);
+    }
     if (cacheStyle) {
       return cacheStyle;
     }
@@ -2947,7 +2954,11 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     const style = column?.style;
     cacheStyle = <FullExtendStyle>columnStyleContents.of(
       style,
-      this.theme.bodyStyle,
+      layoutMap.isBottomFrozenRow(row) && this.theme.bottomFrozenStyle
+        ? this.theme.bottomFrozenStyle
+        : layoutMap.isRightFrozenColumn(col) && this.theme.rightFrozenStyle
+        ? this.theme.rightFrozenStyle
+        : this.theme.bodyStyle,
       {
         col,
         row,
@@ -2960,13 +2971,18 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
       this.options.autoWrapText
     );
     if (!isFunction(style)) {
-      this.bodyStyleCache.set(cacheKey, cacheStyle);
+      if (layoutMap.isBottomFrozenRow(row)) {
+        this.bodyBottomStyleCache.set(cacheKey, cacheStyle);
+      } else {
+        this.bodyStyleCache.set(cacheKey, cacheStyle);
+      }
     }
     return cacheStyle;
   }
   clearCellStyleCache() {
     this.headerStyleCache.clear();
     this.bodyStyleCache.clear();
+    this.bodyBottomStyleCache.clear();
   }
   /**
    * 清除行高度缓存对象
@@ -3624,5 +3640,50 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
       col: col + this.rowHeaderLevelCount,
       row: row + this.columnHeaderLevelCount
     };
+  }
+
+  /**
+   * 监听vchart事件
+   * @param type vchart事件类型
+   * @param listener vchart事件监听器
+   * @returns 事件监听器id
+   */
+  onVChartEvent(type: string, callback: AnyFunction): void;
+  onVChartEvent(type: string, query: any, callback: AnyFunction): void;
+  onVChartEvent(type: string, query?: any, callback?: AnyFunction): void {
+    if (!this._chartEventMap[type]) {
+      this._chartEventMap[type] = [];
+    }
+    if (typeof query === 'function') {
+      this._chartEventMap[type].push({ callback: query });
+    } else {
+      this._chartEventMap[type].push({ callback, query });
+    }
+  }
+
+  offVChartEvent(type: string, callback?: AnyFunction): void {
+    // delete this._chartEventMap[type];
+    if (!this._chartEventMap[type]) {
+      return;
+    }
+    if (callback) {
+      this._chartEventMap[type] = this._chartEventMap[type].filter(e => e.callback !== callback);
+    } else {
+      this._chartEventMap[type] = [];
+    }
+  }
+  /** 给activeChartInstance逐个绑定chart用户监听事件 */
+  _bindChartEvent(activeChartInstance: any) {
+    if (activeChartInstance) {
+      for (const key in this._chartEventMap) {
+        (this._chartEventMap[key] || []).forEach(e => {
+          if (e.query) {
+            activeChartInstance.on(key, e.query, e.callback);
+          } else {
+            activeChartInstance.on(key, e.callback);
+          }
+        });
+      }
+    }
   }
 }