+++ packages/vtable/src/ts-types/list-table/define/basic-define.ts
@@ -54,6 +54,7 @@ export interface IBasicHeaderDefine {
    *  是否禁用调整列宽,如果是转置表格或者是透视表的指标是行方向指定 那该配置不生效
    */
   disableColumnResize?: boolean;
+  headerEditor?: string | IEditor | ((args: BaseCellInfo & { table: BaseTableAPI }) => string | IEditor);
 }
 
 export interface IBasicColumnBodyDefine {